thumb, retries, etc

+++ livecam.py
@@ -56,6 +56,13 @@ def main():
             'source': 'https://www.youtube.com/channel/UCgdHxnHSXvcAi4PaMIY1Ltg', 
             'uploader': 'SHIBUYA COMMUNITY NEWS', 
         }, 
+        'spain1': {
+            'originalurl': 'https://www.youtube.com/watch?v=2qAQ1DUFWhY', 
+            'licenseurl': 'https://creativecommons.org/licenses/by/3.0/', 
+            'tags': ["madrid", "spain", "Sierra de Guadarrama", "birds", "aves", "bird feeder", "comedero"], 
+            'source': 'https://www.youtube.com/channel/UCg6OElsAGYmkuSySkZpaNGw', 
+            'uploader': 'SEOBirdLife - Sociedad Española de Ornitología', 
+        }, 
     }
     
     livecam = ""
@@ -63,12 +70,15 @@ def main():
         livecam = sys.argv[1] in livecams.keys() and sys.argv[1] or ''
         
     if livecam:        
+        maxretries = 5
         originalurl = livecams[livecam]['originalurl']
         destfile = "livecam-%s-%s.mp4" % (livecam, todayandhour)
+        destthumbfile = "livecam-%s-%s-thumb.jpg" % (livecam, today)
         timeout = 60*60+60 #in seconds, 1 hour + 1 minute
         #timeout = 20 #in seconds, 20 seconds for tests
         os.system("timeout -s 15 %ss python youtube-dl %s -o %s" % (timeout, originalurl, destfile))
         os.system("mv %s.part %s" % (destfile, destfile))
+        os.system("python youtube-dl %s --skip-download --write-thumbnail -o %s" % (originalurl, destthumbfile))
         
         itemid = "livecam-%s-%s" % (livecam, today)
         itemtags = commontags + livecams[livecam]['tags'] + [livecam]
@@ -85,14 +95,27 @@ def main():
             'date': today_, 
             'year': year, 
             'subject': '; '.join(itemtags), 
-            'licenseurl': livecams[livecam]['licenseurl'], 
+            'licenseurl': 'licenseurl' in livecams[livecam] and livecams[livecam]['licenseurl'] or '', 
             'originalurl': originalurl,
         }
-        item = get_item(itemid)
-        item.upload(destfile, metadata=md, access_key=accesskey, secret_key=secretkey, verbose=True, queue_derive=False)
-        item.modify_metadata(md)
-        print('You can find it in https://archive.org/details/%s' % (itemid))
-        os.remove(destfile)
+        retries = 1
+        uploaded = False
+        while retries <= maxretries:
+            try:
+                item = get_item(itemid)
+                item.upload(files=[destfile, destthumbfile], metadata=md, access_key=accesskey, secret_key=secretkey, verbose=True, queue_derive=False)
+                item.modify_metadata(md)
+                uploaded = True
+                break
+            except:
+                pass
+            time.sleep(600*retries)
+            retries += 1
+        if uploaded:
+            print('You can find it in https://archive.org/details/%s' % (itemid))
+            os.remove(destfile)
+        else:
+            print('Error while uploading. Upload it manually later.')
     else:
         livecamslist = list(livecams.keys())
         livecamslist.sort()
